Analysis of how accurate is decision of belonging to the same family based on last name only - how many last names are there that look like they are actually not the same family (not in the same class, not same deck, not same embarked port). It's important since one of the most important features is % survived per family.  If it's not the same family, then relying on this % is not accurate. Conclusion: seems there are quite a few like that. e see that 8 families appear in test group that also different at least by 2 out of 3 parameters of Pclass, Embarked, Deck. There is a good chance that for these familities we have a lot of bad predictions since we are relying on % survival rate, but either it's not accurate already in training set, or it's not the same family in test set.
Best (and probably would improve accuracy of the model) would be to use the information above not to combine into the same family last names that don't share the same Deck / Embarked port / Pclass, or differ in 2 out of 3 parameters. This would probably make the feature % survival rate of the family available for less families, but stronger.

However, since the feature of % survival rate is very complicated already, decided not to do it currently.